fix eslint and preitter Relates #1
