+++ client/.eslintrc.js
@@ -1,23 +1,22 @@
 module.exports = {
+  root: true,
   env: {
     browser: true,
     es2021: true,
   },
   extends: [
     'plugin:react/recommended',
-     "some-other-config-you-use",
-     "prettier",
     'airbnb',
+    'prettier',
+    'plugin:react/jsx-runtime',
   ],
-  overrides: [
-  ],
+  overrides: [],
   parserOptions: {
     ecmaVersion: 'latest',
     sourceType: 'module',
   },
-  plugins: [
-    'react',
-  ],
+  plugins: ['react', 'prettier'],
   rules: {
+    'prettier/prettier': 'error',
   },
 };
