+++ client/src/App.jsx
@@ -1,0 +1,7 @@
+import './App.css';
+
+function App() {
+  return <div className="App">Hello from the other side - adele</div>;
+}
+
+export default App;
