+++ package.json
@@ -4,6 +4,8 @@
   "description": "e-commerce app",
   "main": "index.js",
   "scripts": {
+    "build": "cd client && npm run build",
+    "client:start": "cd client && npm start",
     "start": "cross-env NODE_ENV=production node src/index",
     "dev": "cross-env NODE_ENV=development nodemon src/index",
     "test": "cross-env NODE_ENV=test jest",
@@ -28,9 +30,10 @@
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "jest": "^29.0.3",
     "nodemon": "^2.0.20",
-    "prettier": "^2.7.1",
+    "prettier": "2.7.1",
     "stylelint-config-prettier": "^9.0.3",
     "supertest": "^6.2.4"
   }
